[FLINK-14050][yarn] Refactor YarnClusterDescriptor inheritance
