+++ flink-yarn-tests/src/test/java/org/apache/flink/yarn/util/NonDeployingYarnClusterDescriptor.java
@@ -23,7 +23,7 @@ import org.apache.flink.client.program.ClusterClient;
 import org.apache.flink.configuration.Configuration;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.util.Preconditions;
-import org.apache.flink.yarn.AbstractYarnClusterDescriptor;
+import org.apache.flink.yarn.YarnClusterDescriptor;
 
 import org.apache.hadoop.yarn.api.records.ApplicationId;
 import org.apache.hadoop.yarn.api.records.ApplicationReport;
@@ -31,9 +31,9 @@ import org.apache.hadoop.yarn.client.api.YarnClient;
 import org.apache.hadoop.yarn.conf.YarnConfiguration;
 
 /**
- * Dummy {@link AbstractYarnClusterDescriptor} without an actual deployment for tests.
+ * Dummy {@link YarnClusterDescriptor} without an actual deployment for tests.
  */
-public class NonDeployingYarnClusterDescriptor extends AbstractYarnClusterDescriptor {
+public class NonDeployingYarnClusterDescriptor extends YarnClusterDescriptor {
 
 	private final ClusterClient<ApplicationId> clusterClient;
 
@@ -45,19 +45,17 @@ public class NonDeployingYarnClusterDescriptor extends AbstractYarnClusterDescri
 			ClusterClient<ApplicationId> clusterClient) {
 		super(flinkConfiguration, yarnConfiguration, configurationDirectory, yarnClient, true);
 
-		//noinspection unchecked
 		this.clusterClient = Preconditions.checkNotNull(clusterClient);
 	}
 
 	@Override
 	public String getClusterDescription() {
-		// return parent.getClusterDescription();
 		return "NonDeployingYarnClusterDescriptor";
 	}
 
 	@Override
 	protected ClusterClient<ApplicationId> createYarnClusterClient(
-			AbstractYarnClusterDescriptor descriptor,
+			YarnClusterDescriptor descriptor,
 			int numberTaskManagers,
 			int slotsPerTaskManager,
 			ApplicationReport report,
@@ -78,21 +76,9 @@ public class NonDeployingYarnClusterDescriptor extends AbstractYarnClusterDescri
 
 	@Override
 	public ClusterClient<ApplicationId> deployJobCluster(
-			ClusterSpecification clusterSpecification, JobGraph jobGraph, boolean detached) {
+			ClusterSpecification clusterSpecification,
+			JobGraph jobGraph,
+			boolean detached) {
 		return clusterClient;
-	}
-
-	@Override
-	public void killCluster(ApplicationId clusterId) {
-	}
-
-	@Override
-	protected String getYarnSessionClusterEntrypoint() {
-		throw new UnsupportedOperationException("Not needed in test.");
-	}
-
-	@Override
-	protected String getYarnJobClusterEntrypoint() {
-		throw new UnsupportedOperationException("Not needed in test.");
 	}
 }
