+++ flink-yarn/src/test/java/org/apache/flink/yarn/AbstractYarnClusterTest.java
@@ -18,12 +18,8 @@
 
 package org.apache.flink.yarn;
 
-import org.apache.flink.client.deployment.ClusterDeploymentException;
 import org.apache.flink.client.deployment.ClusterRetrieveException;
-import org.apache.flink.client.deployment.ClusterSpecification;
-import org.apache.flink.client.program.ClusterClient;
 import org.apache.flink.configuration.Configuration;
-import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.util.Preconditions;
 import org.apache.flink.util.TestLogger;
 
@@ -45,7 +41,7 @@ import java.util.Collections;
 import java.util.Map;
 
 /**
- * Tests for the {@link AbstractYarnClusterDescriptor}.
+ * Tests for the {@link YarnClusterDescriptor}.
  */
 public class AbstractYarnClusterTest extends TestLogger {
 
@@ -68,7 +64,7 @@ public class AbstractYarnClusterTest extends TestLogger {
 		yarnClient.init(yarnConfiguration);
 		yarnClient.start();
 
-		final TestingAbstractYarnClusterDescriptor clusterDescriptor = new TestingAbstractYarnClusterDescriptor(
+		final YarnClusterDescriptor clusterDescriptor = new YarnClusterDescriptor(
 			new Configuration(),
 			yarnConfiguration,
 			temporaryFolder.newFolder().getAbsolutePath(),
@@ -124,38 +120,6 @@ public class AbstractYarnClusterTest extends TestLogger {
 			} else {
 				return super.getApplicationReport(appId);
 			}
-		}
-	}
-
-	private static final class TestingAbstractYarnClusterDescriptor extends AbstractYarnClusterDescriptor {
-
-		private TestingAbstractYarnClusterDescriptor(
-				Configuration flinkConfiguration,
-				YarnConfiguration yarnConfiguration,
-				String configurationDirectory,
-				YarnClient yarnClient,
-				boolean sharedYarnClient) {
-			super(flinkConfiguration, yarnConfiguration, configurationDirectory, yarnClient, sharedYarnClient);
-		}
-
-		@Override
-		protected String getYarnSessionClusterEntrypoint() {
-			throw new UnsupportedOperationException("Not needed for testing");
-		}
-
-		@Override
-		protected String getYarnJobClusterEntrypoint() {
-			throw new UnsupportedOperationException("Not needed for testing");
-		}
-
-		@Override
-		protected ClusterClient<ApplicationId> createYarnClusterClient(AbstractYarnClusterDescriptor descriptor, int numberTaskManagers, int slotsPerTaskManager, ApplicationReport report, Configuration flinkConfiguration, boolean perJobCluster) throws Exception {
-			throw new UnsupportedOperationException("Not needed for testing");
-		}
-
-		@Override
-		public ClusterClient<ApplicationId> deployJobCluster(ClusterSpecification clusterSpecification, JobGraph jobGraph, boolean detached) throws ClusterDeploymentException {
-			throw new UnsupportedOperationException("Not needed for testing");
 		}
 	}
 }
